+++ packages/typespec-rust/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.36.1 (unreleased)
 
+### Features Added
+
+* Define a `pub(crate)` constant for `api-version` to use in hand-authored `Default` implementation on client options.
+
 ### Other Changes
 
 * Updated to the latest tsp toolset.